fix(test): correct the type of index in loop to fit `size_type` of `Vector`

+++ test/test_vector.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #define CATCH_CONFIG_MAIN
 #include "Vector.hpp"
 #include "catch2/catch.hpp"
@@ -7,7 +8,7 @@ TEST_CASE("Vector initialize", "[Vector]") {
     dsa::Vector<int> v2(v1);
     // TODO: `==` operator is not implemented
     REQUIRE(v1.size() == v2.size());
-    for (size_t i = 0; i < v1.size(); ++i) {
+    for (int64_t i = 0; i < v1.size(); ++i) {
         REQUIRE(v1[i] == v2[i]);
     }
 }
